+++ R/add_manual_locked_project_constraints.R
@@ -1,15 +1,15 @@
 #' @include internal.R ProjectProblem-class.R
 NULL
 
-#' Add manually specified locked constraints for actions
+#' Add manually specified locked constraints for projects
 #'
 #' Add constraints to a project prioritization problem to ensure
-#' that particular actions are selected, or not selected, for funding
+#' that particular projects are selected, or not selected, for funding
 #' by the solution. This function offers
-#' more fine-grained control than the [add_locked_in_action_constraints()]
-#' and [add_locked_out_action_constraints()] functions.
+#' more fine-grained control than the [add_locked_in_project_constraints()]
+#' and [add_locked_out_project_constraints()] functions.
 #'
-#' @inheritParams add_locked_in_action_constraints
+#' @inheritParams add_locked_in_project_constraints
 #'
 #' @param locked `data.frame` or [tibble::tibble()] object. See
 #' the Details section for more information.
@@ -19,16 +19,16 @@ NULL
 #'
 #' \describe{
 #'
-#' \item{`"action"`}{`character` values with action names.}
+#' \item{`"project"`}{`character` values with project names.}
 #'
 #' \item{`"status"`}{
-#' `numeric` values indicating if actions should
+#' `numeric` values indicating if projects should
 #'  be selected for funding (with a value of 1) or not (with a value of zero).
 #' }
 #'
 #' }
 #'
-#' @inherit add_locked_in_action_constraints return seealso
+#' @inherit add_locked_in_project_constraints return seealso
 #'
 #' @family constraints
 #'
@@ -39,7 +39,7 @@ NULL
 #'
 #' # create data frame with locked statuses
 #' locked_data <- data.frame(
-#'   action = sim_actions$name[1:2],
+#'   project = sim_projects$name[1:2],
 #'   status = c(0, 1)
 #' )
 #'
@@ -54,7 +54,7 @@ NULL
 #'     "name", "success", "name", "cost", "name"
 #'   ) %>%
 #'   add_max_wtd_sum_objective(budget = 500) %>%
-#'   add_manual_locked_action_constraints(locked_data) %>%
+#'   add_manual_locked_project_constraints(locked_data) %>%
 #'   add_binary_decisions()
 #'
 #' # print problem
@@ -67,24 +67,24 @@ NULL
 #' print(s)
 #' }
 #'
-#' @name add_manual_locked_action_constraints
+#' @name add_manual_locked_project_constraints
 #'
-#' @exportMethod add_manual_locked_action_constraints
+#' @exportMethod add_manual_locked_project_constraints
 #'
-#' @aliases add_manual_locked_action_constraints,ProjectProblem,data.frame-method add_manual_locked_action_constraints,ProjectProblem,tbl_df-method
+#' @aliases add_manual_locked_project_constraints,ProjectProblem,data.frame-method add_manual_locked_project_constraints,ProjectProblem,tbl_df-method
 #'
 #' @export
 methods::setGeneric(
-  "add_manual_locked_action_constraints",
+  "add_manual_locked_project_constraints",
   signature = methods::signature("x", "locked"),
-  function(x, locked) standardGeneric("add_manual_locked_action_constraints")
+  function(x, locked) standardGeneric("add_manual_locked_project_constraints")
 )
 
-#' @name add_manual_locked_action_constraints
-#' @usage \S4method{add_manual_locked_action_constraints}{ProjectProblem,data.frame}(x, locked)
-#' @rdname add_manual_locked_action_constraints
+#' @name add_manual_locked_project_constraints
+#' @usage \S4method{add_manual_locked_project_constraints}{ProjectProblem,data.frame}(x, locked)
+#' @rdname add_manual_locked_project_constraints
 methods::setMethod(
-  "add_manual_locked_action_constraints",
+  "add_manual_locked_project_constraints",
   methods::signature("ProjectProblem", "data.frame"),
   function(x, locked) {
     # assert valid arguments
@@ -93,15 +93,15 @@ methods::setMethod(
       inherits(locked, "data.frame")
     )
     # add constraints
-    add_manual_locked_action_constraints(x, tibble::as_tibble(locked))
+    add_manual_locked_project_constraints(x, tibble::as_tibble(locked))
   }
 )
 
-#' @name add_manual_locked_action_constraints
-#' @usage \S4method{add_manual_locked_action_constraints}{ProjectProblem,tbl_df}(x, locked)
-#' @rdname add_manual_locked_action_constraints
+#' @name add_manual_locked_project_constraints
+#' @usage \S4method{add_manual_locked_project_constraints}{ProjectProblem,tbl_df}(x, locked)
+#' @rdname add_manual_locked_project_constraints
 methods::setMethod(
-  "add_manual_locked_action_constraints",
+  "add_manual_locked_project_constraints",
   methods::signature("ProjectProblem", "tbl_df"),
   function(x, locked) {
     # assert arguments are valid
@@ -109,11 +109,10 @@ methods::setMethod(
       inherits(x, "ProjectProblem"),
       inherits(locked, "tbl_df"),
       nrow(locked) > 0,
-      assertthat::has_name(locked, "action"),
-      inherits(locked$action, c("character", "factor")),
-      assertthat::noNA(locked$action),
-      all(locked$action %in%
-        as.character(x$action_names())),
+      assertthat::has_name(locked, "project"),
+      inherits(locked$project, c("character", "factor")),
+      assertthat::noNA(locked$project),
+      all(locked$project %in% as.character(x$project_names())),
       assertthat::has_name(locked, "status"),
       is.numeric(locked$status),
       all(locked$status %in% c(0, 1)),
@@ -122,13 +121,13 @@ methods::setMethod(
     # set attributes
     if (all(locked$status == 1)) {
       class_name <- "LockedInConstraint"
-      constraint_name <- "locked in actions"
+      constraint_name <- "locked in projects"
     } else if (all(!locked$status == 0)) {
       class_name <- "LockedOutConstraint"
-      constraint_name <- "locked out actions"
+      constraint_name <- "locked out projects"
     } else {
       class_name <- "LockedManualConstraint"
-      constraint_name <- "manually locked actions"
+      constraint_name <- "manually locked projects"
     }
     # add constraints
     x$add_constraint(
@@ -137,7 +136,7 @@ methods::setMethod(
         inherit = Constraint,
         public = list(
           name = constraint_name,
-          data = list(action_names = x$action_names(), locked = locked),
+          data = list(project_names = x$project_names(), locked = locked),
           apply = function(x, y) {
             assertthat::assert_that(
               inherits(x, "OptimizationProblem"),
@@ -145,9 +144,9 @@ methods::setMethod(
             )
             d <- self$get_data("locked")
             invisible(
-              rcpp_apply_locked_action_constraints(
+              rcpp_apply_locked_project_constraints(
                 x$ptr,
-                match(d$action, self$data$action_names),
+                match(d$project, self$data$project_names),
                 as.integer(d$status)
               )
             )