+++ src/rcpp_apply_locked_action_constraints.cpp
@@ -5,10 +5,12 @@
 bool rcpp_apply_locked_action_constraints(SEXP x,
                                           Rcpp::IntegerVector actions,
                                           Rcpp::NumericVector status) {
+  // initialization
   Rcpp::XPtr<OPTIMIZATIONPROBLEM> ptr = Rcpp::as<Rcpp::XPtr<OPTIMIZATIONPROBLEM>>(x);
+  std::size_t n = actions.size();
 
   // apply constraints
-  for (std::size_t i = 0; i < static_cast<std::size_t>(actions.size()); ++i) {
+  for (std::size_t i = 0; i < n; ++i) {
     if (status[i] < 0.5) {
       ptr->_ub[actions[i] - 1] = 0.0;
     } else {