+++ tests/testthat/test_add_locked_in_project_constraints.R
@@ -1,6 +1,7 @@
 test_that("integer (compile)", {
   # create problem
   data(sim_projects, sim_actions, sim_features)
+  sim_projects$locked_in <- c(TRUE, rep(FALSE, nrow(sim_projects) - 1))
   # build problem
   p <-
     problem(
@@ -8,13 +9,13 @@ test_that("integer (compile)", {
       "name", "success", "name", "cost", "name", FALSE
     ) %>%
     add_max_wtd_sum_objective(1) %>%
-    add_locked_in_action_constraints(which(sim_actions$locked_in))
+    add_locked_in_project_constraints(which(sim_projects$locked_in))
   # compile problem
   o <- compile(p)
   # run tests
   expect_equal(
     o$lb(),
-    replace(o$lb(), which(sim_actions$locked_in), 1)
+    replace(o$lb(), nrow(sim_actions) + which(sim_projects$locked_in), 1)
   )
 })
 
@@ -24,6 +25,7 @@ test_that("integer (solve)", {
   skip_if_not(any_solvers_installed())
   # load data
   data(sim_projects, sim_actions, sim_features)
+  sim_projects$locked_in <- c(TRUE, rep(FALSE, nrow(sim_projects) - 1))
   # build problem
   p <-
     problem(
@@ -31,18 +33,19 @@ test_that("integer (solve)", {
       "name", "success", "name", "cost", "name", FALSE
     ) %>%
     add_max_wtd_sum_objective(1e+5) %>%
-    add_locked_in_action_constraints(which(sim_actions$locked_in))
+    add_locked_in_project_constraints(which(sim_projects$locked_in))
   # solve problem
   s <- solve(p)
   # run tests
-  for (i in sim_actions$name[sim_actions$locked_in]) {
-    expect_equal(s[[i]], 1)
+  for (i in sim_projects$name[sim_projects$locked_in]) {
+    expect_equal(s[[i]], TRUE)
   }
 })
 
 test_that("integer (invalid arguments)", {
   # load data
   data(sim_projects, sim_actions, sim_features)
+  sim_projects$locked_in <- c(TRUE, rep(FALSE, nrow(sim_projects) - 1))
   # build problem
   p <-
     problem(
@@ -52,22 +55,23 @@ test_that("integer (invalid arguments)", {
     add_max_wtd_sum_objective(1)
   # run tests
   expect_error({
-    add_locked_in_action_constraints(p, -1)
+    add_locked_in_project_constraints(p, -1)
   })
   expect_error({
-    add_locked_in_action_constraints(p, 0)
+    add_locked_in_project_constraints(p, 0)
   })
   expect_error({
-    add_locked_in_action_constraints(p, nrow(sim_actions) + 1)
+    add_locked_in_project_constraints(p, nrow(sim_actions) + 1)
   })
   expect_error({
-    add_locked_in_action_constraints(p, NA_real_)
+    add_locked_in_project_constraints(p, NA_real_)
   })
 })
 
 test_that("logical (compile)", {
   # load data
   data(sim_projects, sim_actions, sim_features)
+  sim_projects$locked_in <- c(TRUE, rep(FALSE, nrow(sim_projects) - 1))
   # build problem
   p <-
     problem(
@@ -75,19 +79,20 @@ test_that("logical (compile)", {
       "name", "success", "name", "cost", "name", FALSE
     ) %>%
     add_max_wtd_sum_objective(1) %>%
-    add_locked_in_action_constraints(sim_actions$locked_in)
+    add_locked_in_project_constraints(sim_projects$locked_in)
   # compile problem
   o <- compile(p)
   # run tests
   expect_equal(
     o$lb(),
-    replace(o$lb(), which(sim_actions$locked_in), 1)
+    replace(o$lb(), nrow(sim_actions) + which(sim_projects$locked_in), 1)
   )
 })
 
 test_that("logical (invalid arguments)", {
   # load data
   data(sim_projects, sim_actions, sim_features)
+  sim_projects$locked_in <- c(TRUE, rep(FALSE, nrow(sim_projects) - 1))
   # build problem
   p <-
     problem(
@@ -97,13 +102,13 @@ test_that("logical (invalid arguments)", {
     add_max_wtd_sum_objective(1)
   # run tests
   expect_error({
-    add_locked_in_action_constraints(p, FALSE)
+    add_locked_in_project_constraints(p, FALSE)
   })
   expect_error({
-    add_locked_in_action_constraints(p, c(TRUE, TRUE))
+    add_locked_in_project_constraints(p, c(TRUE, TRUE))
   })
   expect_error({
-    add_locked_in_action_constraints(p, NA)
+    add_locked_in_project_constraints(p, NA)
   })
 })
 
@@ -113,6 +118,7 @@ test_that("logical (solve)", {
   skip_if_not(any_solvers_installed())
   # load data
   data(sim_projects, sim_actions, sim_features)
+  sim_projects$locked_in <- c(TRUE, rep(FALSE, nrow(sim_projects) - 1))
   # build problem
   p <-
     problem(
@@ -120,18 +126,19 @@ test_that("logical (solve)", {
       "name", "success", "name", "cost", "name", FALSE
     ) %>%
     add_max_wtd_sum_objective(1e+5) %>%
-    add_locked_in_action_constraints(sim_actions$locked_in)
+    add_locked_in_project_constraints(sim_projects$locked_in)
   # solve problem
   s <- solve(p)
   # run tests
-  for (i in sim_actions$name[sim_actions$locked_in]) {
-    expect_equal(s[[i]], 1)
+  for (i in sim_projects$name[sim_projects$locked_in]) {
+    expect_equal(s[[i]], TRUE)
   }
 })
 
 test_that("character (compile)", {
   # load data
   data(sim_projects, sim_actions, sim_features)
+  sim_projects$locked_in <- c(TRUE, rep(FALSE, nrow(sim_projects) - 1))
   # build problem
   p <-
     problem(
@@ -139,19 +146,20 @@ test_that("character (compile)", {
       "name", "success", "name", "cost", "name", FALSE
     ) %>%
     add_max_wtd_sum_objective(1) %>%
-    add_locked_in_action_constraints("locked_in")
+    add_locked_in_project_constraints("locked_in")
   # compile problem
   o <- compile(p)
   # run tests
   expect_equal(
     o$lb(),
-    replace(o$lb(), which(sim_actions$locked_in), 1)
+    replace(o$lb(), nrow(sim_actions) + which(sim_projects$locked_in), 1)
   )
 })
 
 test_that("logical (invalid arguments)", {
   # load data
   data(sim_projects, sim_actions, sim_features)
+  sim_projects$locked_in <- c(TRUE, rep(FALSE, nrow(sim_projects) - 1))
   # build problem
   p <-
     problem(
@@ -161,13 +169,13 @@ test_that("logical (invalid arguments)", {
     add_max_wtd_sum_objective(1)
   # run tests
   expect_error({
-    add_locked_in_action_constraints(p, "name")
+    add_locked_in_project_constraints(p, "name")
   })
   expect_error({
-    add_locked_in_action_constraints(p, "column_that_doesnt_exist")
+    add_locked_in_project_constraints(p, "column_that_doesnt_exist")
   })
   expect_error({
-    add_locked_in_action_constraints(p, NA_character_)
+    add_locked_in_project_constraints(p, NA_character_)
   })
 })
 
@@ -177,6 +185,7 @@ test_that("character (solve)", {
   skip_if_not(any_solvers_installed())
   # load data
   data(sim_projects, sim_actions, sim_features)
+  sim_projects$locked_in <- c(TRUE, rep(FALSE, nrow(sim_projects) - 1))
   # build problem
   p <-
     problem(
@@ -184,11 +193,11 @@ test_that("character (solve)", {
       "name", "success", "name", "cost", "name", FALSE
     ) %>%
     add_max_wtd_sum_objective(1e+5) %>%
-    add_locked_in_action_constraints("locked_in")
+    add_locked_in_project_constraints("locked_in")
   # solve problem
   s <- solve(p)
   # run tests
-  for (i in sim_actions$name[sim_actions$locked_in]) {
-    expect_equal(s[[i]], 1)
+  for (i in sim_projects$name[sim_projects$locked_in]) {
+    expect_equal(s[[i]], TRUE)
   }
 })